+++ src/MazeHunter.Core/Enemies/EnemyKind.cs
@@ -1,0 +1,7 @@
+namespace MazeHunter.Core.Enemies;
+
+public enum EnemyKind : byte
+{
+    Drifter
+}
+